🚚 api: Move certification-course route to session-management context
